pyproject.toml
[tool.poetry]
name = "python-lessons"
version = "0.1.0"
description = ""
authors = ["Oleg Averkin <olaverkin@lohika.com>"]

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^7.4.3"
pytest-playwright = "^0.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
